+++ CHANGELOG.md
@@ -14,6 +14,7 @@ speaking, considered to be the player's experience.
 
 - Show an icon for units which are outside of their general's command range and
   so do not receive combat bonuses and are losing loyaly.
+- Display unit leadership effect status in the info pane.
 
 ### Fixed